Refactor token storage to follow project structure and configuration
- Use UseUnoPersistenceCache() for Uno-specific persistent storage
- Split Request/Handler into separate files following project structure
- Move cache configuration to appsettings.json instead of attributes
- Update CLAUDE.md with detailed project structure
- Configure 30-day token persistence via configuration

Structure:
- /Mediator/Requests/Authentication/ - Request objects
- /Mediator/Handlers/Authentication/ - Handler implementations
- Cache config in appsettings.json under Mediator.Cache
