+++ src/UnoApp/UnoApp/Mediator/Requests/Authentication/ClearTokenRequest.cs
@@ -1,0 +1,5 @@
+using Shiny.Mediator;
+
+namespace UnoApp.Mediator.Requests.Authentication;
+
+public record ClearTokenRequest : IRequest;
