+++ src/UnoApp/UnoApp/appsettings.json
@@ -4,6 +4,12 @@
       "Debug": true,
       // * works here - all http contracts will pick up a baseURI
       "UnoApp.ApiClient.*": "https://api.themeparks.wiki/v1"
+    },
+    "Cache": {
+      "UnoApp.Mediator.Requests.Authentication.GetStoredTokenRequest": {
+        "AbsoluteExpirationSeconds": 2592000,
+        "StorageType": "Persistent"
+      }
     }
   },
     "AppConfig": {
